Implemented searchProduct endpoint and worked on product part of store to get image url, size and color of product(#1wn444v)
