+++ src/locales/en.json
@@ -5,17 +5,19 @@
   "Download orders":"Download orders",
   "In Stock": "{count} In Stock",
   "Last brokered":"Last brokered",
+  "Location inventory":"Location inventory",
   "Logout":"Logout",
   "Orders":"Orders",
+  "Ordered on":"Ordered on",
   "PO arrival date":"PO arrival date",
   "Promise date":"Promise date",
   "Ready to create an app?": "Ready to create an app?",
   "Settings":"Settings",
+  "Shipping from":"Shipping from",
+  "Shipping method":"Shipping method",
   "Size":"Size",
   "Start with Ionic": "Start with Ionic",
   "Store":"Store",
   "store name":"store name",
   "UI Components":"UI Components"
-
-  
 }
