+++ src/main.ts
@@ -57,7 +57,7 @@ app.config.globalProperties.$filters = {
       externalId = externalIdentificationSplit[1] ? externalIdentificationSplit[1] : '';
     }
     return externalId;
-  },
+  }, 
   getFeature(featureHierarchy: any, featureKey: string) {
     let  featureValue = ''
     if (featureHierarchy) {
@@ -77,7 +77,7 @@ app.config.globalProperties.$filters = {
       }
     }
     return customerLoyalty;
-  }
+  },
 }
