+++ src/router/index.ts
@@ -1,7 +1,7 @@
 import { createRouter, createWebHistory } from '@ionic/vue-router';
 import { RouteRecordRaw } from 'vue-router';
 import Home from '@/views/Home.vue'
-import Order from '@/views/Order.vue'
+import Order from '@/views/Orders.vue'
 // import ProductInventory from '@/views/ProductInventory.vue'
 import Login from '@/views/Login.vue'
 import Settings from "@/views/Settings.vue"
