+++ src/store/modules/product/ProductState.ts
@@ -1,6 +1,8 @@
 export default interface ProductState {
+  cached: any;  
   products: {
     list: any;
     total: number;
   }
+  productsInformation: any;
 }
