+++ src/store/modules/product/actions.ts
@@ -40,11 +40,47 @@ const actions: ActionTree<ProductState, RootState> = {
       // Remove added loader only when new query and not the infinite scroll
       if (payload.viewIndex === 0) emitter.emit("dismissLoader");
     } catch(error){
-      console.log(error)
       showToast(translate("Something went wrong"));
     }
     // TODO Handle specific error
     return resp;
+  },
+
+  // Will fetch product information
+  async fetchProducts ( { commit, state }, { productIds }) {
+    const cachedProductIds = Object.keys(state.cached);
+    const productIdFilter= productIds.reduce((filter: string, productId: any) => {
+      if (filter !== '') filter += ' OR '
+      if (cachedProductIds.includes(productId)) {
+        return filter;
+      } else {
+        return filter += productId;
+      }
+    }, '');
+
+    if (productIdFilter === '') return;
+    const resp = await ProductService.fetchProducts({
+      "filters": ['productId: (' + productIdFilter + ')']
+    })
+    if (resp.status === 200 && !hasError(resp)) {
+      const products = resp.data.response.docs;
+      if (resp.data) commit(types.PRODUCT_ADD_TO_CACHED_MULTIPLE, { products });
+    }
+    return resp;
+  },
+
+  // Get product related information
+  async getProductsInformation  ( context , { products }) {
+    // To remove redundant value Set is used
+    let productIds: any = new Set();
+    products.groups.forEach((product: any) => {
+      productIds.add(product.groupValue);
+    });
+    // Converted to list as methods like reduce not supported
+    productIds = [...productIds]
+    if (productIds.length) {
+      this.dispatch('product/fetchProducts', { productIds })
+    }
   },
 }
