+++ src/store/modules/product/getters.ts
@@ -3,6 +3,13 @@ import ProductState from "./ProductState";
 import RootState from "../../RootState";
 
 const getters: GetterTree<ProductState, RootState> = {
+  getCached (state) {
+    return state.cached;
+  },
+  getProduct: (state) => (productId: string) => {
+    // Returning empty object so that it doesn't breaks the UI
+    return state.cached[productId] ? state.cached[productId] : {};
+  },
   getSearchProducts(state) {
     return state.products.list;
   },
