+++ src/store/modules/product/mutation-types.ts
@@ -1,2 +1,3 @@
 export const SN_PRODUCT = 'product'
 export const PRODUCT_SEARCH_UPDATED = SN_PRODUCT + '/SEARCH_UPDATED'
+export const PRODUCT_ADD_TO_CACHED_MULTIPLE = SN_PRODUCT + '/ADD_TO_CACHED_MULTIPLE'
